+++ src/main/resources/application.yml
@@ -2,14 +2,15 @@ app:
   modes:
     enabled: INGRESS, PROCESSOR_MQTT, PROCESSOR_INFLUX
 
+# Or maybe like that?
+#      enabled: [ingress, processor_mqtt, processor_influx]
+
 messaging:
   retry:
     message-sender-retry-count: 5
     message-sender-retry-initial-back-off-interval: 1100
     message-sender-back-off-multiplier: 1.5
     message-retry-intervals: 10000,30000,300000
-    application-modes:
-      enabled: [ingress, processor_mqtt, processor_influx]
 
 mqtt:
   enabled: false
